ref - remove dead code, consolidate things, clean
