+++ backend/src/models/ExportJob.js
@@ -116,10 +116,6 @@ const exportJobSchema = new mongoose.Schema(
       type: String,
       required: false
     },
-    lastProcessedId: {
-      type: mongoose.Schema.Types.ObjectId,
-      required: false
-    }
   },
   {
     timestamps: true
